+++ price-feeder/cmd/price-feeder.go
@@ -13,18 +13,19 @@ import (
 	"time"
 
 	input "github.com/cosmos/cosmos-sdk/client/input"
+	"github.com/mitchellh/mapstructure"
 
 	"github.com/gorilla/mux"
 	"github.com/rs/zerolog"
 	"github.com/spf13/cobra"
 	"golang.org/x/sync/errgroup"
 
+	"github.com/cosmos/cosmos-sdk/telemetry"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/umee-network/umee/price-feeder/config"
 	"github.com/umee-network/umee/price-feeder/oracle"
 	"github.com/umee-network/umee/price-feeder/oracle/client"
 	v1 "github.com/umee-network/umee/price-feeder/router/v1"
-	"github.com/umee-network/umee/price-feeder/telemetry"
 )
 
 const (
@@ -101,7 +102,7 @@ func priceFeederCmdHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	ctx, cancel := context.WithCancel(context.Background())
+	ctx, cancel := context.WithCancel(cmd.Context())
 	g, ctx := errgroup.WithContext(ctx)
 
 	// listen for and trap any OS signal to gracefully shutdown and exit
@@ -164,7 +165,12 @@ func priceFeederCmdHandler(cmd *cobra.Command, args []string) error {
 		endpoints,
 	)
 
-	metrics, err := telemetry.New(cfg.Telemetry)
+	telemetryCfg := telemetry.Config{}
+	err = mapstructure.Decode(cfg.Telemetry, &telemetryCfg)
+	if err != nil {
+		return err
+	}
+	metrics, err := telemetry.New(telemetryCfg)
 	if err != nil {
 		return err
 	}
@@ -230,10 +236,11 @@ func startPriceFeeder(
 
 	srvErrCh := make(chan error, 1)
 	srv := &http.Server{
-		Handler:      rtr,
-		Addr:         cfg.Server.ListenAddr,
-		WriteTimeout: writeTimeout,
-		ReadTimeout:  readTimeout,
+		Handler:           rtr,
+		Addr:              cfg.Server.ListenAddr,
+		WriteTimeout:      writeTimeout,
+		ReadTimeout:       readTimeout,
+		ReadHeaderTimeout: readTimeout,
 	}
 
 	go func() {
@@ -244,7 +251,7 @@ func startPriceFeeder(
 	for {
 		select {
 		case <-ctx.Done():
-			shutdownCtx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
+			shutdownCtx, cancel := context.WithTimeout(ctx, 15*time.Second)
 			defer cancel()
 
 			logger.Info().Str("listen_addr", cfg.Server.ListenAddr).Msg("shutting down price-feeder server...")